refactor: Extract Updater interface and add BgUpdater/SimpleUpdater implementations
